+++ odoo/local-src/sale_discount_program/models/program.py
@@ -115,23 +115,33 @@ class Program(models.Model):
             else:
                 program.voucher_amount = False
 
+    def _get_action_values_for_voucher_amount(self, product_add_price):
+        return {
+            'type_action': 'product_add',
+            'product_add_id': self.env.ref(
+                'sale_discount_program.product_voucher'
+            ).id,
+            'product_add_force_price': True,
+            'product_add_price': product_add_price,
+            'allow_negative_total': False,
+        }
+
     def _inverse_voucher_amount(self):
         for program in self:
             if program.voucher_amount:
+                # Let the get product_add_price value here,
+                # because unlink will be drop it
                 product_add_price = -1 * program.voucher_amount
                 if program.action_ids:
                     program.action_ids.unlink()
 
+                action_values = program._get_action_values_for_voucher_amount(
+                    product_add_price
+                )
                 program.write({
-                    "action_ids": [(0, False, {
-                        'type_action': 'product_add',
-                        'product_add_id': self.env.ref(
-                            'sale_discount_program.product_voucher'
-                        ).id,
-                        'product_add_force_price': True,
-                        'product_add_price': product_add_price,
-                        'allow_negative_total': False,
-                    })]
+                    "action_ids": [
+                        (0, False, action_values)
+                    ]
                 })
 
     @api.depends('expiration_date', 'nb_use', 'max_use')